Revert changes to Core InfoTable data structure
Using NameId instead of Text in the Core InfoTable broke the Core REPL
and use of NameIds in Core is problematic (see #1564).

The textual representation of Identifier indices in the Info table can
now be chosen when calling `runInfoTableBuilder`. In the Internal to
Core translation we format the nameId field and in the source to Core
translation we use the nameText field.