Restore macOS CI build/test
The key to get this working was to use the installation of LLVM that's
already present on the macOS agent, available at `$(brew --prefix
llvm@14)/bin`. The LLVM installed by the `install-llvm-action` appears
to be broken for macOS.

Other changes:

- The libclang_rt.builtins-wasm32-wasi installation is not required for
newer LLVM versions, so I've removed this.
- Use `runner.os` instead of `matrix.os` to detect OS. `runner.os` is
more stable, it doesn't change when the os image version is updated.