Add transformation group synonyms in `core read -t` (#1868)
This pr adds the option to pass the arguments `pipeline-eval`,
`pipeline-geb`, `pipeline-stripped` to `core read -t`. These will
desugar as expected to the corresponding sequences of transformations.

This pr also removes the default `pipeline-eval` if the user does not
provide an explicit transformation. Instead, no transformations will be
applied by default. To have the same behavour as now, the option `-t
pipeline-eval` will need to be passed