CI: Ignore errors linux typecheck / format examples step
There's an error with this step on linux only, but we cannot see it
because of the `-s` flag on the makefile.

This commit removes the `-s` flag so we can diagnose the problem, but
also temporarily ignores the error to avoid blocking other PRs.

NB: This step passes on macOS.