Let-folding after lifting (#1955)
Now that lifting preserves the types, we can fold the lets afterwards.
This is a cheap optimisation that should always be done. In the WASM
pipeline it avoids allocating closures for constant functions assigned
to let-variables. In the GEB pipeline it reduces the size of the term
and the number of variables.