Fix recurse interpretation in walkDirRel
A function passed to RecurseFilter controls which subdirectories
should be traversed in the directory walk.

The first argument indicates if the candidate subdirectory contains a juvix.yaml
file and the second argument is the path of the candidate subdirectory.

Previously the first argument was calculated from the files in the
current directory - which was incorrect.

The callers of walkDirRel are also fixed. The intended behaviour is to
avoid traversing into hidden directories and into juvix
sub-projects (i.e subdirectories containing a juvix.yaml file).