Fix extra whitespace in text rendering of indented empty lines (#2101)
The prettyprinter library takes care avoid adding whitespace to empty
lines when it is rendering indented text.

See:

https://github.com/quchen/prettyprinter/blob/7e32c010ecfaa510c0cbf6ffea04bc7db95d4fcb/prettyprinter/src/Prettyprinter/Internal.hs#L1999

However it only does this for unannotated text.

In our code we were stripping annotations from renderings within
`toTextStream` but we must remove the annotations before calling
`layoutPretty` to get the proper handling of whitespace with
indentations. That's what this PR does.